Cockpit+Gauge: Show if robot code exists
