+++ common/Cargo.toml
@@ -16,3 +16,4 @@ ws = "0.9.2"
 bus = "2.4.0"
 serde_json = "1.0.95"
 systemstat = "0.2.3"
+home = "0.5.4"
