+++ common/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod config;
 pub mod logging;
 pub mod messages;
-pub mod service_info;
 pub mod system_info;
